chore(host): cleanup hint parsing
