+++ book/src/sdk/pipeline/intro.md
@@ -95,7 +95,7 @@ the [`PipelineBuilder`][builder] to instantiate a [`DerivationPipeline`][dp].
 use std::sync::Arc;
 use op_alloy_protocol::BlockInfo;
 use op_alloy_genesis::RollupConfig;
-use superchain_derive::*;
+use hilo_providers_alloy::*;
 
 // Use a default rollup config.
 let rollup_config = Arc::new(RollupConfig::default());
